refactor: 도메인 설계 수정
- article에서 userRef, user에서 articleRef 를 갖도록 수정
- 과연 이게 양방향 관계라고 할 수 없는지는 의문..

+++ src/main/java/spring/board/demo/service/UserService.java
@@ -1,6 +1,7 @@
 package spring.board.demo.service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
@@ -16,6 +17,7 @@ import spring.board.demo.domain.user.dto.UserCreateResponse;
 import spring.board.demo.domain.user.dto.UserResponse;
 import spring.board.demo.domain.user.dto.UserUpdateRequest;
 import spring.board.demo.exception.NotFoundUserException;
+import spring.board.demo.exception.NotMatchUserIdException;
 
 @Service
 public class UserService {
@@ -58,7 +60,10 @@ public class UserService {
         userRepository.save(user);
     }
 
-    public void delete(Long id) {
+    public void delete(User user, Long id) {
+        if (!Objects.equals(user.getId(), id)) {
+            throw new NotMatchUserIdException();
+        }
         userRepository.deleteById(id);
     }
